# Conflicts:
#	src/main/java/com/stockmarket/www/controller/ManageStockController.java
#	src/main/java/com/stockmarket/www/dao/mybatis/mapper/InterestStocksDao.xml
#	src/main/java/com/stockmarket/www/service/basic/BasicAssetTrendService.java
#	src/main/java/com/stockmarket/www/service/basic/BasicHaveStockService.java
#	src/main/java/com/stockmarket/www/service/basic/BasicInterestStockService.java
#	src/main/webapp/resource/js/managestocks/holding_list.js

+++ src/main/java/com/stockmarket/www/entity/CaptureMemo.java
@@ -173,7 +173,7 @@ public class CaptureMemo {
 		return foreignInvestors;
 	}
 
-	public void setForeignInvestors(float foreignInvestors) {
+	public void setForeignInvestors(double foreignInvestors) {
 		this.foreignInvestors = foreignInvestors;
 	}
